+++ src/app-routes.tsx
@@ -4,9 +4,8 @@ import { Route, Routes } from 'react-router-dom';
 
 import { appRoutesObj } from './app.paths';
 import TasksPage from './pages/user/tasks-page/tasks-page';
-import HomePage from './pages/user/home/home-page';
+import HomePage from './pages/user/dashboard-page/dashboard-page';
 
-const ExamplePage = React.lazy(() => import('./pages/user/example-page'));
 const LoginPage = React.lazy(
   () => import('./pages/user/login-page/login-page'),
 );
@@ -14,7 +13,6 @@ const RegisterPage = React.lazy(
   () => import('./pages/user/register-page/register-page'),
 );
 
-const NotFoundPage = React.lazy(() => import('./pages/404'));
 
 const withSuspense = (WrappedComponent: JSX.Element) => {
   return (
@@ -37,11 +35,6 @@ export function AppRouting() {
           path={appRoutesObj.getRegisterPath()}
           element={withSuspense(<RegisterPage />)}
         />
-        <Route
-          key="examplePage"
-          path={appRoutesObj.getExamplePagePath()}
-          element={withSuspense(<ExamplePage />)}
-        />
 
         <Route
           key="homePage"
@@ -61,11 +54,6 @@ export function AppRouting() {
           element={withSuspense(<TasksPage />)}
         />
 
-        <Route
-          key="notDefined"
-          path="*"
-          element={withSuspense(<NotFoundPage />)}
-        />
       </Routes>
     </Suspense>
   );
