--- src/components/common/tasks/new-task.tsx
@@ -1,51 +1,0 @@
-import { useState } from 'react';
-
-export default function NewTaskForm({ onCreate, onClose }: any) {
-  const [title, setTitle] = useState('');
-  const [description, setDescription] = useState('');
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    onCreate({ title, description });
-  };
-
-  return (
-    <form onSubmit={handleSubmit} className="space-y-4">
-      <div>
-        <label className="block text-sm font-medium text-gray-700">Title</label>
-        <input
-          value={title}
-          onChange={(e) => setTitle(e.target.value)}
-          className="mt-1 w-full rounded-md border border-gray-300 p-2"
-          required
-        />
-      </div>
-      <div>
-        <label className="block text-sm font-medium text-gray-700">
-          Description
-        </label>
-        <textarea
-          value={description}
-          onChange={(e) => setDescription(e.target.value)}
-          className="mt-1 w-full rounded-md border border-gray-300 p-2"
-        />
-      </div>
-
-      <div className="flex justify-end gap-2">
-        <button
-          type="button"
-          onClick={onClose}
-          className="text-gray-500 hover:underline"
-        >
-          Cancel
-        </button>
-        <button
-          type="submit"
-          className="rounded-md bg-purple-600 px-4 py-2 text-white hover:bg-purple-700"
-        >
-          Create
-        </button>
-      </div>
-    </form>
-  );
-}
