+++ src/components/dashboard/TaskProgressChart.tsx
