+++ src/components/dashboard/WeeklyActivityChart.tsx
