--- src/containers/README.md
@@ -1,2 +1,0 @@
-This folder contains all app containers using by pages components.
-the container should naming with name of related page and prefixed with `-container`.
