--- src/containers/user/example-container.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function ExampleContainer() {
-  return <div></div>;
-}
