--- src/core/enums/example.enum.ts
@@ -1,28 +1,0 @@
-export enum example {
-  example = 'example',
-}
-
-export enum LanguagesEnum {
-  EN = 'en',
-  AR = 'ar',
-}
-
-export enum ColorsEnum {
-  Primary = 'primary',
-}
-export enum SizesEnum {
-  Small = 'sm',
-  Medium = 'md',
-  Large = 'lg',
-}
-
-export enum DirectionEnum {
-  RTL = 'rtl',
-  LTR = 'ltr',
-}
-export enum InputStylesEnum {
-  STYLE1 = 'style1',
-  STYLE2 = 'style2',
-  STYLE3 = 'style3',
-  STYLE4 = 'style4',
-}
