+++ src/index.tsx
@@ -1,5 +1,4 @@
 import './App.css';
-import './core/configs/i18n';
 
 import React from 'react';
 import { createRoot } from 'react-dom/client';
