--- src/pages/404.tsx
@@ -1,3 +1,0 @@
-export default function Custom404(): JSX.Element {
-  return <h1>Not Found</h1>;
-}
