--- src/pages/500.tsx
@@ -1,3 +1,0 @@
-export default function Custom500(): JSX.Element {
-  return <h1 className="text-center">Server Error - 500</h1>;
-}
