--- src/pages/admin/readme.md
@@ -1,9 +1,0 @@
-### this folder contains all admin pages, and must be authenticated.
-
-## every pages should be wrapped Admin component
-
-```
-<Admin>
-componenet content
-</Admin>
-```
