--- src/pages/user/example-page.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import ExampleContainer from '../../containers/user/example-container';
-
-export default function ExamplePage() {
-  return (
-    <div>
-      <ExampleContainer />
-    </div>
-  );
-}
